pyproject.toml
[project]
name = "bachelor-project"
version = "0.1.0"
description = ""
authors = [
    {name = "amr.saleh@i-hq.tech"}
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "numpy (>=2.3.3,<3.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.33,<0.4.0)",
    "langgraph (>=0.6.8,<0.7.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "langchain-community (>=0.3.30,<0.4.0)",
    "pypdf (>=6.1.1,<7.0.0)",
    "docarray (>=0.41.0,<0.42.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.115.0,<0.116.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "copilotkit (>=0.1.65,<0.2.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "langgraph-checkpoint (==2.1.2)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "tavily-python (>=0.7.12,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
